Make PassthroughArgsCompiler test-only
Now that all callables use one of the generic arguments compilers or
one of their own, the CallableMetadataBuilder can start requiring all
callables to explicitly specify an arguments compiler.

This should let us drop PassthroughArgsCompiler entirely, but the
compiler tests themselves still require something to validate argument
processing.  So, instead of removing this, make it test-only for the
compiler.
